refactor: MovieInformationStackView 분리
